Add Medicine page added

+++ app/src/main/java/com/example/coagusearch/Class/Medicine.kt
@@ -3,8 +3,8 @@ package com.example.coagusearch.Class
 class Medicine {
     var name:String?=null
     var frequency: MedicineFrequency?=null
-    var dosage:Double?=null
-    constructor(name:String, frequency: MedicineFrequency, dosage: Double){
+    var dosage:Int?=null
+    constructor(name:String, frequency: MedicineFrequency, dosage: Int){
         this.name=name
         this.frequency=frequency
         this.dosage=dosage
